Ref: extracting schema helpers (#1441)
Just moving some stuff
