+++ src/common-helpers.ts
@@ -70,8 +70,6 @@ export const getInput = (
     );
 };
 
-export const isValidDate = (date: Date): boolean => !isNaN(date.getTime());
-
 export const makeErrorFromAnything = (subject: unknown): Error =>
   subject instanceof Error
     ? subject
@@ -305,13 +303,3 @@ export const tryToTransform = <T>(
     return undefined;
   }
 };
-
-// obtaining the private helper type from Zod
-export type ErrMessage = Exclude<
-  Parameters<typeof z.ZodString.prototype.email>[0],
-  undefined
->;
-
-// the copy of the private Zod errorUtil.errToObj
-export const errToObj = (message: ErrMessage | undefined) =>
-  typeof message === "string" ? { message } : message || {};
