+++ src/date-in-schema.ts
@@ -8,7 +8,7 @@ import {
   ZodTypeDef,
   addIssueToContext,
 } from "zod";
-import { isValidDate } from "./common-helpers";
+import { isValidDate } from "./schema-helpers";
 
 // simple regex for ISO date, supports the following formats:
 // 2021-01-01T00:00:00.000Z
