+++ src/date-out-schema.ts
@@ -8,7 +8,7 @@ import {
   ZodTypeDef,
   addIssueToContext,
 } from "zod";
-import { isValidDate } from "./common-helpers";
+import { isValidDate } from "./schema-helpers";
 
 const zodDateOutKind = "ZodDateOut";
