+++ src/schema-helpers.ts
@@ -1,0 +1,17 @@
+import { z } from "zod";
+
+// obtaining the private helper type from Zod
+export type ErrMessage = Exclude<
+  Parameters<typeof z.ZodString.prototype.email>[0],
+  undefined
+>;
+
+// the copy of the private Zod errorUtil.errToObj
+export const errToObj = (message: ErrMessage | undefined) =>
+  typeof message === "string" ? { message } : message || {};
+
+export const isValidDate = (date: Date): boolean => !isNaN(date.getTime());
+
+export const bufferSchema = z.custom<Buffer>((subject) =>
+  Buffer.isBuffer(subject),
+);
